Windows Fixes Patch up stat and io Fix template shared_from_this() snafu undef #ERROR from the GDI, we do this be undefining it anywhere a header touches windows internally

+++ src/flair/internal/services/IAsyncIOService.h
@@ -51,7 +51,7 @@ namespace services {
       {
          bool exists;
          bool isDirectory;
-         uint64_t size;
+         size_t size;
          std::time_t created;
          std::time_t modified;
       };
